Move cu files to detail
