+++ src/QMCWaveFunctions/CMakeLists.txt
@@ -96,23 +96,23 @@ IF(OHMMS_DIM MATCHES 3)
 
   IF(QMC_CUDA)
     SET(WFSSRCS_CUDA
-      Fermion/determinant_update.cu
-      Fermion/delayed_update.cu
+      detail/CUDA_legacy/determinant_update.cu
+      detail/CUDA_legacy/delayed_update.cu
       Fermion/DiracDeterminantCUDA.cpp
       Fermion/SlaterDetCUDA.cpp
       TrialWaveFunction_CUDA.cpp
       )
     SET(WFSSRCS_CUDA ${WFSSRCS_CUDA}
-      Jastrow/BsplineJastrowCuda.cu
-      Jastrow/BsplineJastrowCudaPBC.cu
+      detail/CUDA_legacy/BsplineJastrowCuda.cu
+      detail/CUDA_legacy/BsplineJastrowCudaPBC.cu
       Jastrow/OneBodyJastrowOrbitalBspline.cpp
       Jastrow/TwoBodyJastrowOrbitalBspline.cpp
       )
     IF(HAVE_EINSPLINE)
       SET(WFSSRCS_CUDA ${WFSSRCS_CUDA}
         EinsplineSetCuda.cpp
-        PhaseFactors.cu
-        AtomicOrbitalCuda.cu
+        detail/CUDA_legacy/PhaseFactors.cu
+        detail/CUDA_legacy/AtomicOrbitalCuda.cu
         )
     ENDIF(HAVE_EINSPLINE)
   ENDIF(QMC_CUDA)
@@ -198,8 +198,8 @@ SET(FERMION_SRCS ${FERMION_SRCS}
 
 IF(ENABLE_CUDA)
   SET(WFSSRCS_CUDA ${WFSSRCS_CUDA}
-      Fermion/delayed_update_helper.cu
-      Fermion/matrix_update_helper.cu
+      detail/CUDA/delayed_update_helper.cu
+      detail/CUDA/matrix_update_helper.cu
       )
 ENDIF(ENABLE_CUDA)
