+++ src/QMCWaveFunctions/Fermion/DelayedUpdateCUDA.h
@@ -17,7 +17,7 @@
 #include "CUDA/CUDAallocator.hpp"
 #include <CUDA/cuBLAS.hpp>
 #include <CUDA/cusolver.hpp>
-#include "QMCWaveFunctions/Fermion/delayed_update_helper.h"
+#include "QMCWaveFunctions/detail/CUDA/delayed_update_helper.h"
 #include "QMCWaveFunctions/Fermion/DiracMatrix.h"
 #include <cuda_runtime_api.h>
 #include "CUDA/cudaError.h"
