+++ src/QMCWaveFunctions/Fermion/DiracDeterminantCUDA.cpp
@@ -20,8 +20,8 @@
 
 #include "DiracDeterminantCUDA.h"
 #include "CUDA_legacy/cuda_inverse.h"
-#include "QMCWaveFunctions/Fermion/determinant_update.h"
-#include "QMCWaveFunctions/Fermion/delayed_update.h"
+#include "QMCWaveFunctions/detail/CUDA_legacy/determinant_update.h"
+#include "QMCWaveFunctions/detail/CUDA_legacy/delayed_update.h"
 #include "Numerics/DeterminantOperators.h"
 #include <unistd.h>
