+++ src/QMCWaveFunctions/Fermion/DiracDeterminantCUDA.h
@@ -23,7 +23,7 @@
 #include <typeinfo>
 #include "QMCWaveFunctions/Fermion/DiracDeterminantBase.h"
 #include "QMCWaveFunctions/SPOSet.h"
-#include "QMCWaveFunctions/Fermion/determinant_update.h"
+#include "QMCWaveFunctions/detail/CUDA_legacy/determinant_update.h"
 #include "Utilities/NewTimer.h"
 
 namespace qmcplusplus
