+++ src/QMCWaveFunctions/detail/CUDA/subtractOne.cuh
