+++ src/QMCWaveFunctions/detail/CUDA_legacy/AtomicOrbitalCuda.h
