+++ src/QMCWaveFunctions/detail/CUDA_legacy/BsplineJastrowCuda.cu
