+++ src/QMCWaveFunctions/detail/CUDA_legacy/BsplineJastrowCuda.h
