+++ src/QMCWaveFunctions/detail/CUDA_legacy/BsplineJastrowCudaPBC.h
