+++ src/QMCWaveFunctions/detail/CUDA_legacy/NLjobGPU.h
