+++ src/QMCWaveFunctions/detail/CUDA_legacy/PhaseFactors.cu
@@ -14,6 +14,7 @@
 //////////////////////////////////////////////////////////////////////////////////////
     
     
+#include "PhaseFactors.h"
 #include <assert.h>
 #include <thrust/complex.h>
