+++ src/QMCWaveFunctions/detail/CUDA_legacy/delayed_update.h
