+++ src/QMCWaveFunctions/detail/CUDA_legacy/determinant_update.h
